remove one subcommand level
