+++ pkg/cmd/kre/runtime/list.go
@@ -1,4 +1,4 @@
-package list
+package runtime
 
 import (
 	"fmt"
