+++ pkg/cmd/kre/runtime/runtime.go
@@ -5,7 +5,6 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/konstellation-io/kli/cmdutil"
-	"github.com/konstellation-io/kli/pkg/cmd/kre/runtime/list"
 )
 
 // NewRuntimeCmd creates a new command to handle 'runtime' subcommands.
@@ -21,7 +20,7 @@ func NewRuntimeCmd(f cmdutil.CmdFactory) *cobra.Command {
 	cmd.PersistentFlags().StringP("server", "s", f.Config().DefaultServer, "KRE server to use")
 
 	cmd.AddCommand(
-		list.NewListCmd(f),
+		NewListCmd(f),
 	)
 
 	return cmd
