+++ pkg/cmd/kre/version/list.go
@@ -1,4 +1,4 @@
-package list
+package version
 
 import (
 	"fmt"
