+++ pkg/cmd/kre/version/start.go
@@ -1,17 +1,18 @@
-package unpublish
+package version
 
 import (
 	"github.com/konstellation-io/kli/cmdutil"
 	"github.com/spf13/cobra"
 )
 
-// NewUnpublishCmd unpublishes an existing version.
-func NewUnpublishCmd(f cmdutil.CmdFactory) *cobra.Command {
+// NewStartCmd starts an existing version.
+func NewStartCmd(f cmdutil.CmdFactory) *cobra.Command {
 	log := f.Logger()
 	cmd := &cobra.Command{
-		Use:   "unpublish",
-		Args:  cmdutil.ComposeArgsCheck(cmdutil.CheckServerFlag, cobra.ExactArgs(1)),
-		Short: "Unpublish a version",
+		Use:     "start",
+		Aliases: []string{"start"},
+		Args:    cmdutil.ComposeArgsCheck(cmdutil.CheckServerFlag, cobra.ExactArgs(1)),
+		Short:   "Start a version",
 		RunE: func(cmd *cobra.Command, args []string) error {
 			s, _ := cmd.Flags().GetString("server")
 			c, err := f.KreClient(s)
@@ -25,17 +26,17 @@ func NewUnpublishCmd(f cmdutil.CmdFactory) *cobra.Command {
 			}
 
 			versionID := args[0]
-			err = c.Version().Unpublish(versionID, comment)
+			err = c.Version().Start(versionID, comment)
 			if err != nil {
 				return err
 			}
 
-			log.Success("Unpublishing version '%s'.", versionID)
+			log.Success("Starting version '%s'.", versionID)
 			return nil
 		},
 	}
 
-	cmd.Flags().StringP("message", "m", "", "Adds audit message to unpublish")
+	cmd.Flags().StringP("message", "m", "", "Adds audit message to start")
 	_ = cmd.MarkPersistentFlagRequired("message")
 
 	return cmd
