+++ pkg/cmd/kre/version/unpublish.go
@@ -1,18 +1,17 @@
-package stop
+package version
 
 import (
 	"github.com/konstellation-io/kli/cmdutil"
 	"github.com/spf13/cobra"
 )
 
-// NewStopCmd stop an existing version.
-func NewStopCmd(f cmdutil.CmdFactory) *cobra.Command {
+// NewUnpublishCmd unpublishes an existing version.
+func NewUnpublishCmd(f cmdutil.CmdFactory) *cobra.Command {
 	log := f.Logger()
 	cmd := &cobra.Command{
-		Use:     "stop",
-		Aliases: []string{"stop"},
-		Args:    cmdutil.ComposeArgsCheck(cmdutil.CheckServerFlag, cobra.ExactArgs(1)),
-		Short:   "Stop a version",
+		Use:   "unpublish",
+		Args:  cmdutil.ComposeArgsCheck(cmdutil.CheckServerFlag, cobra.ExactArgs(1)),
+		Short: "Unpublish a version",
 		RunE: func(cmd *cobra.Command, args []string) error {
 			s, _ := cmd.Flags().GetString("server")
 			c, err := f.KreClient(s)
@@ -26,17 +25,17 @@ func NewStopCmd(f cmdutil.CmdFactory) *cobra.Command {
 			}
 
 			versionID := args[0]
-			err = c.Version().Stop(versionID, comment)
+			err = c.Version().Unpublish(versionID, comment)
 			if err != nil {
 				return err
 			}
 
-			log.Success("Stopping version '%s'.", versionID)
+			log.Success("Unpublishing version '%s'.", versionID)
 			return nil
 		},
 	}
 
-	cmd.Flags().StringP("message", "m", "", "Adds audit message to stop")
+	cmd.Flags().StringP("message", "m", "", "Adds audit message to unpublish")
 	_ = cmd.MarkPersistentFlagRequired("message")
 
 	return cmd
