+++ pkg/cmd/kre/version/version.go
@@ -3,11 +3,6 @@ package version
 import (
 	"github.com/MakeNowJust/heredoc"
 	"github.com/konstellation-io/kli/cmdutil"
-	"github.com/konstellation-io/kli/pkg/cmd/kre/version/list"
-	"github.com/konstellation-io/kli/pkg/cmd/kre/version/publish"
-	"github.com/konstellation-io/kli/pkg/cmd/kre/version/start"
-	"github.com/konstellation-io/kli/pkg/cmd/kre/version/stop"
-	"github.com/konstellation-io/kli/pkg/cmd/kre/version/unpublish"
 	"github.com/spf13/cobra"
 )
 
@@ -26,11 +21,11 @@ func NewVersionCmd(f cmdutil.CmdFactory) *cobra.Command {
 	_ = cmd.MarkPersistentFlagRequired("runtime")
 
 	cmd.AddCommand(
-		list.NewListCmd(f),
-		start.NewStartCmd(f),
-		stop.NewStopCmd(f),
-		publish.NewPublishCmd(f),
-		unpublish.NewUnpublishCmd(f),
+		NewListCmd(f),
+		NewStartCmd(f),
+		NewStopCmd(f),
+		NewPublishCmd(f),
+		NewUnpublishCmd(f),
 	)
 
 	return cmd
